+++ cx-portal/src/utils/downloadUtils.ts
@@ -1,0 +1,7 @@
+export function download(file: Blob, fileType: string, fileName: string) {
+  const url = URL.createObjectURL(file)
+  const a = document.createElement('a')
+  a.href = url
+  a.download = fileName
+  a.click()
+}